// modules/apps/headless/headless-admin-site/headless-admin-site-client-js/src/models/FragmentConfigurationFieldValue.ts
/**
 * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */


/**
 * @author Rubén Pulido
 * @generated
 */

	/**
	* The value of a fragment configuration field.
	*/
	export class FragmentConfigurationFieldValue {
			"type"?: 'Category' | 'Checkbox' | 'Collection' | 'ColorPalette' | 'ColorPicker' | 'Item' | 'Length' | 'NavigationMenu' | 'Select' | 'Text' | 'URL' | 'Video';

		static "discriminator": string | undefined = "type";

	static "attributeTypeMap": Array<{
		baseName: string;
		name: string;
		type: string;
	}> = [
		{
			baseName: "type",
			name: "type",
			type: "'Category' | 'Checkbox' | 'Collection' | 'ColorPalette' | 'ColorPicker' | 'Item' | 'Length' | 'NavigationMenu' | 'Select' | 'Text' | 'URL' | 'Video'",
		},
		];

		static getAttributeTypeMap() {
				return FragmentConfigurationFieldValue.attributeTypeMap;
		}
	}